+++ internal/provider/provider.go
@@ -58,6 +58,7 @@ func (p *utilsProvider) Functions(ctx context.Context) []func() function.Functio
 		NewYamlDecodeFunction,
 		NewResolveYamlTagsFunction,
 		NewRenderDeviceConfigsFunction,
+    NewVersionCompareFunction,
 	}
 }