+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Retain map ordering in `yaml_merge`
 - Handle merging of scalars and maps in the same list gracefully
 - Fix `yaml_merge` losing string type for values that look like scientific notation, numbers, booleans, or timestamps
+- - Add `version_compare` function
 
 ## 1.0.2